// src/types/partials/InsurancePlanPlan.ts
/* eslint-disable no-unused-vars */
/* eslint-disable no-trailing-spaces */
// This file is auto-generated by generate_classes so do not edit manually

import { TExtension } from '../partials/Extension';
import { TIdentifier } from '../partials/Identifier';
import { TCodeableConcept } from '../partials/CodeableConcept';
import { TReference } from '../partials/Reference';
import { TInsurancePlanGeneralCost } from '../partials/InsurancePlanGeneralCost';
import { TInsurancePlanSpecificCost } from '../partials/InsurancePlanSpecificCost';

export type TInsurancePlanPlan = {
    id?: String;
    extension?: TExtension[];
    modifierExtension?: TExtension[];
    identifier?: TIdentifier[];
    type?: TCodeableConcept;
    coverageArea?: TReference[];
    network?: TReference[];
    generalCost?: TInsurancePlanGeneralCost[];
    specificCost?: TInsurancePlanSpecificCost[];
};